[ProjectTableDemo]: more refactoring.
